requirements.txt
alembic==1.10.1
autopep8==2.0.2
click==8.1.3
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pycodestyle==2.10.0
PyMySQL==1.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
